Remove descending_field parameter
The descending_field parameter was removed beause the strategy for
choosing the next magnetization no longer relies on the direction that
the field is ramping.

Before, if the field was decreasing (descending_field=True), then only
roots with a magnetization less than the previous would be accepted.
However if the field was increasing, only increasing magnetizations would
be accepted.

This constraint was removed by choosing the magnetization that was
closest to the previous one as the new magnetization.