Fix bad results when H_dir and Ku_dir are parallel
When H_dir and Ku_dir were parallel the cross products used to
create the rotation matrix that converts the magnetization back
into x,y,z coordinates would give NaN values.

If H_dir and Ku_dir are detected to be parallel we now find an
orientation vector (either up=(0, 0, 0) or left=(1, 0, 0)) that
is not equal to H_dir and use that to calculate the cross product.

Since there are two choices of orientation vector, H_dir cannot
be parallel to both.